### 2025-01-06 #7630 @editorialbot
[PRE REVIEW]: SweetBean: A declarative language for behavioral experiments with human and artificial participants
**Submitting author:** @younesStrittmatter (<a href="http://orcid.org/0000-0002-3414-2838">Younes Strittmatter</a>)
**Repository:** https://github.com/AutoResearch/sweetbean
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
