### 2025-01-06 #7630 @sappelhoff
@younesStrittmatter do you have suggestions for potential reviewers? Please just have potential [conflicts of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) in mind.

If not, I will consult the automatic suggestions [above](https://github.com/openjournals/joss-reviews/issues/7630#issuecomment-2572902171), or will have a look at the [JOSS reviewer database](https://reviewers.joss.theoj.org/).
