### 2025-01-10 #7630 @younesStrittmatter
Hello @sappelhoff,

Thank you for considering our submission.

I believe Josh de Leeuw (jodeleeuw) would be an ideal choice as a reviewer since this software is heavily based on his jsPsych framework, and he is an expert in online behavioral experiments.

I also think Hause Lin (hauselin) would be great.

If either of them can't review, I think Felix Henninger (felixhenninger) would also be a great fit.
