### 2025-01-17 ba83ae0
bug: fix initialisation of shared variables if they are used in side effects only
