### 2025-01-21 #7703 @editorialbot
[REVIEW]: SweetBean: A declarative language for behavioral experiments with human and artificial participants
**Submitting author:** @younesStrittmatter (<a href="http://orcid.org/0000-0002-3414-2838">Younes Strittmatter</a>)
**Repository:** https://github.com/AutoResearch/sweetbean
**Branch with paper.md** (empty if default branch):
**Version:** v1.3.0
