### 2025-01-22 d350b1a
feat: add first draft of trolley problem
