### 2025-01-22 dc5061f
feat: add preamble and demographic util
