### 2025-02-13 #7703 @jodeleeuw
Hi @younesStrittmatter,

This is really neat work!  Had no issues getting it up and running in Colab. I thought the docs were easy to follow. The text mode and LLM stuff is very cool.

A few thoughts, some of which are relevant to the paper itself and some are just broader thoughts about the relationship between `sweatbean` and `jspsych`.

(1) The paper's statement of need describes part of the need as:

> While platforms like jsPsych offer powerful solutions for running online experiments, they require knowledge of JavaScript, which poses a significant barrier for researchers without programming expertise.

This also seems true of `sweetbean` (just a different language). I think what you are getting at is an idea that the declarative `sweetbean` language is easier for a non-programmer ("`sweetbean` aims to simplify the process of generating online experiments by providing an intuitive declarative language for specifying experiments in Python") because it's less complicated(?) than jsPsych, but I think that the abstraction in `sweetbean` is essentially the same as the abstraction in jsPsych and that the differences are primarily syntax. In fact, I think "By abstracting the details of web development, sweetbean enables researchers to concentrate on the specification of the experiment itself, rather than on its implementation details" is also a pretty apt description of the goals of jsPsych. Does that seem like a fair characterization of the similarity?

That said - I think its super cool that this lets folks write in Python *and* adds important new features on top of that.

(2) Would you mind citing [the more recent jsPsych paper](https://joss.theoj.org/papers/10.21105/joss.05351) which is now the preferred citation for the library? It helps recognize other contributors who are authors on that paper but not the original.

(3) I'm curious about how you are thinking about the relationship between classes in `sweetbean/stimulus` and jsPsych plugins.  One alternative seems to be that there could be a generic wrapper so that any jsPsych plugin could be used from `sweetbean`. Is that something you've thought about? It seems like it would make it easier for folks to jump into using sweetbean if the translation between a jsPsych plugin and `sweetbean` was consistent. I suspect a hangup here is support for the natural language version of each plugin, but this doesn't seem insurmountable...

Really exciting stuff - congrats on the work!
