### 2025-02-15 374b01b
feat: add generic stimuli
