### 2025-02-15 8574e7c
feat: make warning better
