### 2025-02-15 AutoResearch/sweetbean#123 @younesStrittmatter
paper revision
first paper revision based on feedback given here: https://github.com/openjournals/joss-reviews/issues/7703
