+++ paper/paper.bib
@@ -1,12 +1,11 @@
-@article{de2015jspsych,
-  title={jsPsych: A JavaScript library for creating behavioral experiments in a Web browser},
-  author={De Leeuw, Joshua R},
-  journal={Behavior research methods},
-  volume={47},
-  pages={1--12},
-  year={2015},
-  publisher={Springer},
-  doi={10.3758/s13428-014-0458-y}
+@article{de2023jspsych,
+  title={jsPsych: Enabling an open-source collaborative ecosystem of behavioral experiments},
+  author={de Leeuw, Joshua R and Gilbert, Rebecca A and Luchterhandt, Bj{\"o}rn},
+  journal={Journal of Open Source Software},
+  volume={8},
+  number={85},
+  pages={5351},
+  year={2023}
 }
 
 @article{autora2024,
