### 2025-02-15 #7703 @younesStrittmatter
Hello @jodeleeuw,

thanks a lot for the review and the valuable feedback.

---

(1) That is a very fair characterization of the similarity, as evidenced by how easy it was to port the `jsPsych` implementation to `Python`. The two implementations can be described as having a nearly 1:1 mapping. I would add that, in general, researchers in behavioral science are more familiar with `Python` since they use it for data analysis. I adjusted the wording in the paper to this:

>While platforms like `jsPsych` offer powerful solutions for running online experiments, their reliance on *JavaScript* can be a significant barrier for researchers who lack programming expertise in the language. Many researchers in the behavioral sciences, however, are already familiar with *Python* due to its widespread use in data analysis.

Do you think this conveys the message better?

(2) Of course! I changed the citation to reflect that.

(3) This is an excellent idea. I added a (early) version of this for creating HTML-files only (not for language or image sequence generation). There is now a Generic stimulus class where one can use keywords to pass in the exact same arguments as in the jsPsych plugins. For now, one has to manually adjust the generated scripts to "import" the plugins, but the user will be prompted by a warning to do so when they use Generic Stimul.

I am working on a feature that allows the user to implement custom prompts for the natural language version (also in other stimuli then generic). Probably that can then be used to make generic stimuli usable in natural language as well.

---

Thanks again a lot! I really appreciate your suggestions, and I think it is great for enhancing `sweetBean`.  Also, it’s truly exciting to get feedback from the creator of `jsPsych`! As you can probably tell, I’m a big fan of your work!
