### 2025-02-15 e90226a
docs: add hint to pre-release
