### 2025-02-19 #7703 @jodeleeuw
> (3) This is an excellent idea. I added a (early) version of this for creating HTML-files only (not for language or image sequence generation). There is now a Generic stimulus class where one can use keywords to pass in the exact same arguments as in the jsPsych plugins. For now, one has to manually adjust the generated scripts to "import" the plugins, but the user will be prompted by a warning to do so when they use Generic Stimul. The stimulus is included in the pre-release version 1.1.0.

> I am working on a feature that allows the user to implement custom prompts for the natural language version (also in other stimuli then generic). Probably that can then be used to make generic stimuli usable in natural language as well.

Very cool! Looking forward to seeing this. And if there are changes that would make sense to implement upstream to facilitate this kind of work please get in touch.
