### 2025-03-05 AutoResearch/sweetbean#127 @hauselin
I tried Python 3.11.11 and had similar issues. FYI, I'm not running via colab. I'm running the code as a python script in vscode/cursor. It works in your colab notebook, so I don't know what the issue might be.
