### 2025-03-05 #7703 @hauselin
I just tested the library and reviewed the codebase. Overall, sweetbean's approach to experiment generation is interesting. One major issue (which I raised [here](https://github.com/AutoResearch/sweetbean/issues/127)) is that the library generates working index.html files only when I ran the code in colab notebooks online. When I downloaded the notebooks to run locally or write my own code in .py scripts (tested in Python3.11 and 3.13), the library/compiler doesn't seem to work—the resulting index.html doesn't run the experiment. Thoughts as to what's going on (we can resolve it in your repo instead of here).

Here are a few thoughts/comments, in no particular order.

Code/library

1. Compatibility. sweetbean relies on jspsych and python, but I don't see compatibility being documented anywhere. Which jspsych version(s) does it support? Which python versions are compatible with sweetbean?
2. Tests. There are relatively few [tests](https://github.com/AutoResearch/sweetbean/tree/main/tests), which worries me, given the complexities of compiling in general.
3. Potential feature bloat while lacking basic features? sweetbean has lots of potential, but I feel the library could be better if it focused more on ensuring the basic/essential features are available before implementing fancier features (e.g., LLM integration, bandit task examples with LLMs).

Paper

1. jspsych has many extensions. Does sweetbean work with nicely with most of the extensions? If not, sweetbean has similar limitations as psychopy, and how easy/hard is it to make sweetbean work with existing or new jspsych extensions (the limits of sweetbean should be stated in the paper; perhaps there are certain jspsych features that  cannot be or are very hard to recreate with sweetbean)?
2. "By abstracting the details of web development, sweetbean enables researchers to concentrate on the specification of the experiment itself, rather than on its implementation details." On the other hand, this layer of abstraction can also create new, unexpected problems. If a researcher is familiar with Python but not JS, then sweetbean could be very helpful indeed, so I think this is the intended audience of the library? However, in many ways, sweetbean's declarative language uses similar syntax (e.g., nested python dictionaries) that is akin to that in JS, which makes me wonder if it's better to use just learn/use jspsych instead of sweetbean? I guess it boils down to this question: If a behavioral researcher (which I think is the intended audience of the library) had to learn one set of tools/libraries, is the python+sweetbean combo actually preferred over jspsych itself?
3. Related to the above point—the authors said "While platforms like jsPsych offer powerful solutions for running online experiments, they require knowledge of JavaScript, which poses a significant barrier for researchers without programming expertise." Many behavioral researchers are not too familiar with Python (afaik, they use R more than Python), so sweetbean—which uses Python—might not be ideal either? It'd be great if the benefits of sweetbean are highlighted more in the paper; currently, the paper focuses more on what's not great about JS+web development and the problems with PsychoPy, rather than what sweetbean offers.
