### 2025-03-05 8c56edd
doc: add installation/compatibility to documentation
