### 2025-03-05 a290ba0
tests: test now runs each stimulus in browser and checks for errors
