+++ tests/test_stimuli.py
@@ -1,8 +1,11 @@
+import asyncio
 import importlib
 import inspect
 import os
 import pkgutil
 import sys
+
+from pyppeteer import launch
 
 import sweetbean
 from sweetbean import Block, Experiment
@@ -16,6 +19,49 @@ excludes = [
     "_Template_",
     "Generic",
 ]
+
+# Define the excluded stimuli
+EXCLUDES = {
+    "SurveyStimulus",
+    "_BaseStimulus",
+    "_KeyboardResponseStimulus",
+    "_Template_",
+    "Generic",
+}
+
+
+async def test_experiment_in_browser(html_path):
+    """Ensure that generated experiments run correctly in a headless browser."""
+    browser = await launch(headless=True)
+    page = await browser.newPage()
+
+    # Convert the file path to an absolute file:// URL
+    file_url = f"file://{os.path.abspath(html_path)}"
+
+    # List to capture console errors
+    console_errors = []
+
+    async def capture_console(msg):
+        """Capture errors from the browser console."""
+        if msg.type == "error":
+            console_errors.append(msg.text)
+
+    page.on("console", capture_console)  # Listen for console errors
+
+    await page.goto(file_url)
+
+    # Wait for the page to load
+    await page.waitForSelector("body")
+
+    # Ensure the page title is set
+    assert (
+        await page.title() == "My awesome experiment"
+    ), "Experiment did not load correctly!"
+
+    assert not console_errors, f"JavaScript console errors found: {console_errors}"
+
+    await browser.close()
+    print(f"✅ {html_path} loaded successfully in browser.")
 
 
 def test_compile():
@@ -51,6 +97,7 @@ def test_compile():
         assert os.path.exists(
             "basic.html"
         ), f"{stimulus.__name__} failed to generate HTML."
+        asyncio.run(test_experiment_in_browser("basic.html"))
         os.remove("basic.html")
         print(f"{stimulus.__name__} compiled successfully.")
 
