### 2025-03-05 8d2925e
tests: fix headless in github actions
