### 2025-03-05 697dea7
tests: fix headless in github actions
