### 2025-03-05 f95b624
tests: fix headless in github actions
