+++ .github/workflows/test-pytest.yml
@@ -1,6 +1,3 @@
-# This workflow will install Python dependencies, run tests and lint with a variety of Python versions
-# For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
-
 name: Test with py.test
 
 on:
@@ -19,21 +16,43 @@ jobs:
         python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
         os: [ubuntu-latest, macos-latest, windows-latest]
     runs-on: ${{ matrix.os }}
+
     steps:
-    - uses: actions/checkout@v3
-    - uses: actions/setup-python@v4
+    - name: Checkout repository
+      uses: actions/checkout@v3
+
+    - name: Set up Python
+      uses: actions/setup-python@v4
       with:
         python-version: ${{ matrix.python-version }}
         cache: "pip"
-    - name: Install system dependencies for Chromium
-      run: sudo apt-get update && sudo apt-get install -y chromium-browser
 
-    - name: Set up Pyppeteer to use system Chromium
+    # ✅ Install Chromium based on OS
+    - name: Install Chromium (Ubuntu)
+      if: runner.os == 'Linux'
+      run: |
+          sudo apt-get update
+          sudo apt-get install -y chromium-browser || sudo apt-get install -y chromium
+
+    - name: Install Chromium (macOS)
+      if: runner.os == 'macOS'
+      run: brew install chromium
+
+    - name: Install Chromium (Windows)
+      if: runner.os == 'Windows'
+      run: |
+          choco install chromium
+          echo "PYPPETEER_EXECUTABLE_PATH=C:\Program Files\Chromium\Application\chrome.exe" >> $GITHUB_ENV
+
+    # ✅ Set up Pyppeteer to use system Chromium
+    - name: Set up Pyppeteer
       run: |
           mkdir -p ~/.local/share/pyppeteer/local-chromium
-          ln -s /usr/bin/chromium-browser ~/.local/share/pyppeteer/local-chromium/1181205
+          ln -sf $(which chromium-browser || which chromium || which chromium.app/Contents/MacOS/Chromium || echo "C:\Program Files\Chromium\Application\chrome.exe") ~/.local/share/pyppeteer/local-chromium/1181205
+      shell: bash
 
+    # ✅ Run tests
     - name: Run tests
       env:
-        PYPPETEER_EXECUTABLE_PATH: "/usr/bin/chromium-browser"  # ✅ Force Pyppeteer to use system Chromium
+        PYPPETEER_EXECUTABLE_PATH: ${{ env.PYPPETEER_EXECUTABLE_PATH || '/usr/bin/chromium-browser' }}
       run: pytest tests/
