### 2025-03-05 938970b
tests: fix headless in github actions
