### 2025-03-05 166b157
tests: fix headless in github actions
