### 2025-03-05 71b46bd
tests: fix headless in github actions
