### 2025-03-05 3cf57c0
tests: fix headless in github actions
