### 2025-03-05 fc4da5f
tests: fix headless in github actions
