### 2025-03-05 547a20f
tests: fix headless in github actions
