### 2025-03-05 f150feb
tests: fix headless in github actions
