+++ .github/workflows/test-pytest.yml
@@ -27,74 +27,93 @@ jobs:
         python-version: ${{ matrix.python-version }}
         cache: "pip"
 
-    # ✅ Explicitly install dependencies
+    # 1) Install your package (and pytest)
     - name: Install Python dependencies
       run: |
-          python -m pip install --upgrade pip
-          pip install -e .[dev]
-          pip install pytest
+        python -m pip install --upgrade pip
+        pip install -e .[dev]
+        pip install pytest
       shell: bash
 
-    # ✅ Install Chromium based on OS
-    - name: Install Chromium (Ubuntu)
+    # 2) Install Chromium on Linux
+    - name: Install Chromium (Linux)
       if: runner.os == 'Linux'
       run: |
-          sudo apt-get update
-          sudo apt-get install -y chromium-browser || sudo apt-get install -y chromium
+        sudo apt-get update
+        # Try installing 'chromium-browser'; fallback to 'chromium'
+        sudo apt-get install -y chromium-browser || sudo apt-get install -y chromium
 
+    # 3) Install Chromium on macOS
     - name: Install Chromium (macOS)
       if: runner.os == 'macOS'
-      run: |
-          brew install chromium
-          echo "PYPPETEER_EXECUTABLE_PATH=$(which chromium)" >> $GITHUB_ENV
       shell: bash
+      run: |
+        brew install chromium
+        # Set the path for Pyppeteer
+        echo "PYPPETEER_EXECUTABLE_PATH=$(which chromium)" >> $GITHUB_ENV
 
+    # 4) Install Chromium on Windows
     - name: Install Chromium (Windows)
       if: runner.os == 'Windows'
       shell: powershell
       run: |
         choco install chromium -y
-        $chromePath = (Get-Command chrome -ErrorAction SilentlyContinue).Source
-        if ($chromePath) {
-            echo "PYPPETEER_EXECUTABLE_PATH=$chromePath" >> $env:GITHUB_ENV
-        } else {
-            echo "Chromium installation failed or not found!" 
-            exit 1
+        # Check the default install path:
+        $defaultChromeExe = "C:\Program Files\Chromium\Application\chrome.exe"
+        if (Test-Path $defaultChromeExe) {
+          # If it exists, use it
+          echo "PYPPETEER_EXECUTABLE_PATH=$defaultChromeExe" >> $env:GITHUB_ENV
         }
-      
+        else {
+          # Fallback: Try to detect "chrome" via Get-Command
+          $chromePath = (Get-Command chrome -ErrorAction SilentlyContinue).Source
+          if (!$chromePath) {
+            $chromePath = (Get-Command chrome.exe -ErrorAction SilentlyContinue).Source
+          }
+          if (!$chromePath) {
+            Write-Host "Chromium installation failed or not found!"
+            exit 1
+          }
+          echo "PYPPETEER_EXECUTABLE_PATH=$chromePath" >> $env:GITHUB_ENV
+        }
 
-    # ✅ Remove Pyppeteer's auto-downloaded Chromium (forces use of system Chromium)
-    - name: Remove Pyppeteer Chromium Cache
-      run: rm -rf ~/Library/Application\ Support/pyppeteer/local-chromium
+    # 5) Remove Pyppeteer’s auto-downloaded Chromium on macOS
+    - name: Remove Pyppeteer Chromium Cache (macOS)
       if: runner.os == 'macOS'
+      run: rm -rf ~/Library/Application\ Support/pyppeteer/local-chromium
       shell: bash
 
-    # ✅ Ensure Pyppeteer uses the correct Chromium path
-    - name: Set up Pyppeteer
[…]
