### 2025-03-05 5675bb0
tests: fix headless in github actions
