### 2025-03-05 2edf913
tests: fix headless in github actions
