+++ tests/test_stimuli.py
@@ -11,16 +11,7 @@ import sweetbean
 from sweetbean import Block, Experiment
 from sweetbean.stimulus.Stimulus import _BaseStimulus
 
-# Define the excluded stimuli
-excludes = [
-    "SurveyStimulus",
-    "_BaseStimulus",
-    "_KeyboardResponseStimulus",
-    "_Template_",
-    "Generic",
-]
-
-# Define the excluded stimuli
+# Excluded Stimuli
 EXCLUDES = {
     "SurveyStimulus",
     "_BaseStimulus",
@@ -32,83 +23,79 @@ EXCLUDES = {
 
 async def run_experiment_in_browser(html_path):
     """Ensure that generated experiments run correctly in a headless browser."""
-
-    # Force Pyppeteer to use the system-installed Chromium
     executable_path = os.getenv("PYPPETEER_EXECUTABLE_PATH", None)
 
     browser = await launch(
         headless=True,
-        executablePath=executable_path,  # ✅ Explicitly set the Chromium path
+        executablePath=executable_path,
         args=["--no-sandbox", "--disable-setuid-sandbox"],
     )
     page = await browser.newPage()
 
-    # Convert the file path to an absolute file:// URL
     file_url = f"file://{os.path.abspath(html_path)}"
-
-    # List to capture console errors
     console_errors = []
 
     async def capture_console(msg):
-        """Capture errors from the browser console."""
         if msg.type == "error":
             console_errors.append(msg.text)
 
-    page.on("console", capture_console)  # Listen for console errors
-
+    page.on("console", capture_console)
     await page.goto(file_url)
-
-    # Wait for the page to load
     await page.waitForSelector("body")
 
-    # Ensure the page title is set
+    # Check title
     assert (
-        await page.title() == "My awesome experiment"
-    ), "Experiment did not load correctly!"
-
+        await page.title()
+    ) == "My awesome experiment", "Experiment did not load correctly!"
     assert not console_errors, f"JavaScript console errors found: {console_errors}"
 
     await browser.close()
     print(f"✅ {html_path} loaded successfully in browser.")
 
 
-def test_compile():
-    # Dynamically load all modules in the 'sweetbean.stimulus' package
-    stimuli_package = sweetbean.stimulus
+async def compile_all_stimuli():
+    """Gathers all stimuli, compiles them, and runs the browser test in a single event loop."""
 
     # Dynamically load all modules in the 'sweetbean.stimulus' package
+    stimuli_package = sweetbean.stimulus
     for _, module_name, _ in pkgutil.iter_modules(stimuli_package.__path__):
         importlib.import_module(f"sweetbean.stimulus.{module_name}")
 
-    # Gather all subclasses of Stimulus from the loaded modules
+    # Gather all valid stimuli classes
     stimuli_list = []
     for module in sys.modules.values():
         if module and module.__name__.startswith("sweetbean.stimulus"):
             for name, cls in inspect.getmembers(module, inspect.isclass):
-                if issubclass(cls, _BaseStimulus) and cls.__name__ not in excludes:
+                if issubclass(cls, _BaseStimulus) and cls.__name__ not in EXCLUDES:
                     stimuli_list.append(cls)
 
-    # Debugging: Print the found stimuli
-    print()
-    print("Found stimuli:", stimuli_list)
-    print()
+    print(f"\nFound stimuli: {[s.__name__ for s in stimuli_list]}\n")
 
-    for stimulus in stimuli_list:
-        # Log which stimulus is being tested
-        print(f"Testing {stimulus.__name__}...")
+    for stimulus_class in stimuli_list:
+        print(f"Testing {stimulus_class.__name__}...")
 
-        # Test each stimulus
[…]
