### 2025-03-05 9158e2a
tests: fix headless in github actions
