### 2025-03-05 65732ed
tests: fix headless in github actions
