+++ tests/test_stimuli.py
@@ -11,7 +11,7 @@ import sweetbean
 from sweetbean import Block, Experiment
 from sweetbean.stimulus.Stimulus import _BaseStimulus
 
-# Excluded Stimuli
+# Stimuli we don't want to test or are abstract
 EXCLUDES = {
     "SurveyStimulus",
     "_BaseStimulus",
@@ -21,17 +21,32 @@ EXCLUDES = {
 }
 
 
-async def run_experiment_in_browser(html_path):
-    """Ensure that generated experiments run correctly in a headless browser."""
-    executable_path = os.getenv("PYPPETEER_EXECUTABLE_PATH", None)
+# OPTIONAL: If you have a particularly heavy or unstable stimulus (like "Bandit") that
+# you want to skip on CI, uncomment this snippet:
+#
+# SKIP_RESOURCE_INTENSIVE = os.getenv("CI") == "true"
 
+
+async def run_experiment_in_browser(html_path: str):
+    """Launch headless Chromium, load the HTML, check for errors, then close."""
+
+    # Use system Chromium or Pyppeteer-downloaded Chromium
+    executable_path = os.getenv("PYPPETEER_EXECUTABLE_PATH", None)
+    print("Using Chromium path:", executable_path)
+
+    # Launch with recommended flags to prevent random crashes in CI
     browser = await launch(
         headless=True,
         executablePath=executable_path,
-        args=["--no-sandbox", "--disable-setuid-sandbox", "--disable-gpu"],
+        args=[
+            "--no-sandbox",
+            "--disable-setuid-sandbox",
+            "--disable-gpu",
+            "--disable-dev-shm-usage",
+        ],
     )
-    page = await browser.newPage()
 
+    page = await browser.newPage()
     file_url = f"file://{os.path.abspath(html_path)}"
     console_errors = []
 
@@ -39,29 +54,34 @@ async def run_experiment_in_browser(html_path):
         if msg.type == "error":
             console_errors.append(msg.text)
 
+    # Listen for console errors
     page.on("console", capture_console)
-    await page.goto(file_url, options={"timeout": 60000, "waitUntil": "networkidle2"})
-    await page.waitForSelector("body")
 
-    # Check title
-    assert (
-        await page.title()
-    ) == "My awesome experiment", "Experiment did not load correctly!"
+    # Go to the page with a generous timeout and waitUntil=networkidle2
+    await page.goto(file_url, options={"timeout": 60000, "waitUntil": "networkidle2"})
+
+    # Wait a moment if needed to stabilize heavy stimuli
+    await page.waitForSelector("body")
+    await page.waitForTimeout(1000)  # 1-second extra wait for reliability
+
+    # Basic checks
+    title = await page.title()
+    assert title == "My awesome experiment", f"Experiment title mismatch: {title}"
     assert not console_errors, f"JavaScript console errors found: {console_errors}"
 
     await browser.close()
-    print(f"✅ {html_path} loaded successfully in browser.")
+    print(f"✅ {html_path} loaded successfully in browser with no console errors.")
 
 
 async def compile_all_stimuli():
-    """Gathers all stimuli, compiles them, and runs the browser test in a single event loop."""
+    """Aggregate test: compiles each stimulus, runs a headless check in one event loop."""
 
-    # Dynamically load all modules in the 'sweetbean.stimulus' package
+    # Dynamically load all stimuli modules
     stimuli_package = sweetbean.stimulus
     for _, module_name, _ in pkgutil.iter_modules(stimuli_package.__path__):
         importlib.import_module(f"sweetbean.stimulus.{module_name}")
 
-    # Gather all valid stimuli classes
+    # Collect valid stimuli classes
     stimuli_list = []
     for module in sys.modules.values():
[…]
