### 2025-03-05 ed64393
tests: fix headless in github actions
