### 2025-03-05 da13a81
tests: fix headless in github actions
