### 2025-03-05 6cc3f4d
tests: fix headless in github actions
