+++ tests/test_stimuli.py
@@ -1,6 +1,5 @@
 import asyncio
 import importlib
-import inspect
 import os
 import pkgutil
 import sys
@@ -12,7 +11,6 @@ import sweetbean
 from sweetbean import Block, Experiment
 from sweetbean.stimulus.Stimulus import _BaseStimulus
 
-# Stimuli we don't want to test or are abstract
 EXCLUDES = {
     "SurveyStimulus",
     "_BaseStimulus",
@@ -22,20 +20,14 @@ EXCLUDES = {
 }
 
 
-# OPTIONAL: If you have a particularly heavy or unstable stimulus (like "Bandit") that
-# you want to skip on CI, uncomment this snippet:
-#
-# SKIP_RESOURCE_INTENSIVE = os.getenv("CI") == "true"
+# ----- 1. Browser helper function -------------------------------- #
 
 
 async def run_experiment_in_browser(html_path: str):
     """Launch headless Chromium, load the HTML, check for errors, then close."""
-
-    # Use system Chromium or Pyppeteer-downloaded Chromium
     executable_path = os.getenv("PYPPETEER_EXECUTABLE_PATH", None)
     print("Using Chromium path:", executable_path)
 
-    # Launch with recommended flags to prevent random crashes in CI
     browser = await launch(
         headless=True,
         executablePath=executable_path,
@@ -46,84 +38,100 @@ async def run_experiment_in_browser(html_path: str):
             "--disable-dev-shm-usage",
         ],
     )
-
     page = await browser.newPage()
-    file_url = f"file://{os.path.abspath(html_path)}"
+
+    file_url = f"file://{html_path}"
     console_errors = []
 
-    async def capture_console(msg):
+    def capture_console(msg):
         if msg.type == "error":
             console_errors.append(msg.text)
 
-    # Listen for console errors
     page.on("console", capture_console)
 
-    # Go to the page with a generous timeout and waitUntil=networkidle2
+    # Generous timeout, wait for network to be idle
     await page.goto(file_url, options={"timeout": 60000, "waitUntil": "networkidle2"})
 
-    # Wait a moment if needed to stabilize heavy stimuli
+    # If needed, wait for a selector or short sleep
     await page.waitForSelector("body")
-    # await page.waitForTimeout(1000)  # 1-second extra wait for reliability
+    # await asyncio.sleep(1)
 
-    # Basic checks
     title = await page.title()
-    assert title == "My awesome experiment", f"Experiment title mismatch: {title}"
+    assert (
+        title == "My awesome experiment"
+    ), f"Experiment did not load correctly! Title: {title}"
     assert not console_errors, f"JavaScript console errors found: {console_errors}"
 
     await browser.close()
     print(f"✅ {html_path} loaded successfully in browser with no console errors.")
 
 
-async def compile_all_stimuli():
-    """Aggregate test: compiles each stimulus, runs a headless check in one event loop."""
+# ----- 2. Pytest fixture collecting stimuli ------------------------ #
 
-    # Dynamically load all stimuli modules
+
+@pytest.fixture(scope="session")
+def all_stimuli():
+    """Collects all valid stimuli classes from sweetbean.stimulus."""
     stimuli_package = sweetbean.stimulus
     for _, module_name, _ in pkgutil.iter_modules(stimuli_package.__path__):
         importlib.import_module(f"sweetbean.stimulus.{module_name}")
 
-    # Collect valid stimuli classes
     stimuli_list = []
     for module in sys.modules.values():
         if module and module.__name__.startswith("sweetbean.stimulus"):
-            for name, cls in inspect.getmembers(module, inspect.isclass):
-                if issubclass(cls, _BaseStimulus) and cls.__name__ not in EXCLUDES:
[…]
