### 2025-03-05 e8b7fd4
tests: fix headless in github actions
