+++ tests/test_stimuli.py
@@ -20,7 +20,30 @@ EXCLUDES = {
 }
 
 
-# ----- 1. Browser helper function -------------------------------- #
+def get_stimuli_list():
+    """Return all valid stimulus classes from sweetbean.stimulus."""
+    stimuli_package = sweetbean.stimulus
+    # Dynamically import modules
+    for _, module_name, _ in pkgutil.iter_modules(stimuli_package.__path__):
+        importlib.import_module(f"sweetbean.stimulus.{module_name}")
+
+    found = []
+    for module in sys.modules.values():
+        if module and module.__name__.startswith("sweetbean.stimulus"):
+            for name, cls in getattr(module, "__dict__", {}).items():
+                if (
+                    isinstance(cls, type)
+                    and issubclass(cls, _BaseStimulus)
+                    and cls.__name__ not in EXCLUDES
+                ):
+                    found.append(cls)
+
+    print("Collected stimuli:", [cls.__name__ for cls in found])
+    return found
+
+
+# Collect the stimuli classes once
+ALL_STIMULI = get_stimuli_list()
 
 
 async def run_experiment_in_browser(html_path: str):
@@ -40,7 +63,7 @@ async def run_experiment_in_browser(html_path: str):
     )
     page = await browser.newPage()
 
-    file_url = f"file://{html_path}"
+    file_url = f"file://{os.path.abspath(html_path)}"
     console_errors = []
 
     def capture_console(msg):
@@ -49,89 +72,34 @@ async def run_experiment_in_browser(html_path: str):
 
     page.on("console", capture_console)
 
-    # Generous timeout, wait for network to be idle
     await page.goto(file_url, options={"timeout": 60000, "waitUntil": "networkidle2"})
-
-    # If needed, wait for a selector or short sleep
     await page.waitForSelector("body")
-    # await asyncio.sleep(1)
 
     title = await page.title()
-    assert (
-        title == "My awesome experiment"
-    ), f"Experiment did not load correctly! Title: {title}"
-    assert not console_errors, f"JavaScript console errors found: {console_errors}"
+    assert title == "My awesome experiment", f"Title mismatch: {title}"
+    assert not console_errors, f"JS console errors found: {console_errors}"
 
     await browser.close()
-    print(f"✅ {html_path} loaded successfully in browser with no console errors.")
+    print(f"✅ {html_path} loaded successfully.")
 
 
-# ----- 2. Pytest fixture collecting stimuli ------------------------ #
-
-
-@pytest.fixture(scope="session")
-def all_stimuli():
-    """Collects all valid stimuli classes from sweetbean.stimulus."""
-    stimuli_package = sweetbean.stimulus
-    for _, module_name, _ in pkgutil.iter_modules(stimuli_package.__path__):
-        importlib.import_module(f"sweetbean.stimulus.{module_name}")
-
-    stimuli_list = []
-    for module in sys.modules.values():
-        if module and module.__name__.startswith("sweetbean.stimulus"):
-            for name, cls in getattr(module, "__dict__", {}).items():
-                if (
-                    isinstance(cls, type)
-                    and issubclass(cls, _BaseStimulus)
-                    and cls.__name__ not in EXCLUDES
-                ):
-                    stimuli_list.append(cls)
-    return stimuli_list
-
-
-# ----- 3. Parametrized test: one test per stimulus ---------------- #
-
-
-@pytest.mark.parametrize("stimulus_class", [])
+@pytest.mark.parametrize("stimulus_class", ALL_STIMULI, ids=lambda cls: cls.__name__)
 def test_compile_stimulus(stimulus_class):
-    """Compile and test a single stimulus in a headless browser."""
-
[…]
