### 2025-03-05 e43fca5
tests: fix headless in github actions
