### 2025-03-05 87b4466
tests: fix headless in github actions
