### 2025-03-05 b560bf4
tests: fix headless in github actions
