### 2025-03-05 cf1c42c
tests: fix headless in github actions
