### 2025-03-08 #7703 @younesStrittmatter
Hi @hauselin,

Thank you so much for the thorough review and thoughtful feedback! It's incredibly valuable, and I've done my best to address your suggestions. Here's how I've responded to your points:

Bug Fix

The issue you encountered with files not running locally should now be resolved (`v1.0.6`). It was related to the installation process of Transcrypt on your local system. Let me know if it works correctly now!

# Code and Library

1. Compatibility:

Great suggestion! I've updated both the README.md in the repo and added a page to our documentation website [here](https://autoresearch.github.io/sweetbean/installation/) to clearly specify compatibility details, including supported Python and jsPsych versions.

2. Tests:

You're right, more comprehensive tests are essential. Currently, each stimulus is tested for successful compilation, and I've now added an additional automated test that checks if the compiled experiments run without errors directly in the browser. We plan to continuously expand our test coverage going forward.

3. Feature Development:

I completely understand your perspective on prioritizing core functionality. Our team adopts an agile development philosophy, meaning features are often added as research needs arise. Both the LLM integration and bandit task examples were directly driven by our ongoing research projects. I'll certainly keep your advice in mind moving forward, ensuring essential features remain robust and prioritized. Do you have any specific suggestions for core features you'd like to see prioritized, or any areas where you think additional testing would be particularly beneficial? I'd greatly appreciate your insights!

# Paper

1. Compatibility with jsPsych Extensions:

Sweetbean includes a generic "stimulus" type, allowing users to implement jsPsych plugins even if we haven't specifically included them yet. However, this feature is somehow experimental (see review of @jodeleeuw). Additionally, creating new sweetbean stimuli based on jsPsych plugins is relatively straightforward. I encourage users to request stimuli/features they need—we're always happy to add these! I try to emphasize this clearly in the documentation, but I'm open to further suggestions on how to improve this communication.

2./3. Clarifying Sweetbean's Intended Audience and Benefits:

Excellent point. I've added a new section in the paper explicitly outlining when researchers might prefer Sweetbean, jsPsych, or PsychoPy based on their specific needs and background:

---

# When to Use SweetBean, jsPsych, or PsychoPy?

Several tools exist for designing and running behavioral experiments, each with different strengths depending on the research context. Here, we focus on `sweetbean`, `jsPsych`, and `PsychoPy` due to their prominence in web-based and Python-based experiment design.

The choice between `sweetbean`, `jsPsych`, and `PsychoPy` depends on the researcher’s needs, technical expertise, and experimental requirements:

- Use `sweetbean` if you are comfortable with Python and want a declarative workflow that integrates with Python-based tools for experimental design, analysis, and automation. `Sweetbean` is particularly useful for comparing LLM-based experiments with human participants or embedding experiment generation within an automated research pipeline using `sweetpea` and `autora`. It is designed to be accessible and simple, but it is not as feature-rich or general-purpose as `jsPsych` or `PsychoPy`. While existing `jsPsych` plugins may not be directly supported, new plugins can be added based on community demand.

- Use `jsPsych` if you require full flexibility in JavaScript and access to its extensive ecosystem of extensions. If deep customization of experiment logic or fine-tuned control over the execution environment is necessary, `jsPsych` is the better choice. Maintained by a large community, it offers a broader range of plugins and features compared to `sweetbean`. Use `jsPsych` if you need a general-purpose tool for online experiments and are comfortable working with JavaScript.

- Use `PsychoPy` if you prefer a GUI-based approach or need to run experiments offline. While `PsychoPy` supports Python scripting, it is primarily designed for local (rather than web-based) experiments. It is a powerful tool for designing and running a wide range of studies, particularly those requiring precise timing and control over stimulus presentation. Use `PsychoPy` if you want a GUI-based experience, need to run experiments offline, or require high-precision timing control.

---

 SweetBean is part of a set of tools aimed to streamline/automate parts of behavioral research, and I think the best "use case" is when it is used in combination with these tools. However, I hope it is also useful for other researchers that prefer to work with Python and like the declarative, abstract nature of SweetBean.

Your feedback has significantly improved both the library and the clarity of our manuscript. Your points helped clarify Sweetbean's positioning and priorities, which will undoubtedly benefit future users. If there's anything else you'd like me to address or any further insights you'd like to share, I'm very open to it!
