### 2025-03-10 AutoResearch/sweetbean#127 @hauselin
I've installed the latest version v1.0.6, and it's still not working locally if I try to generate the index.html file locally following the procedure above.

Side note: your version numbers need fixing. See [here](https://pypi.org/project/sweetbean/1.0.6/). Your latest version is v1.1.0 which was released in Feb, whereas this version 1.0.6 was released in March.
