### 2025-03-10 #7703 @hauselin
Thanks for updates @younesStrittmatter. I updated my local version to v1.0.6 and it's still not working. See comments [here](https://github.com/AutoResearch/sweetbean/issues/127).
