### 2025-03-10 AutoResearch/sweetbean#127 @hauselin
I have transcrypt in my virtual environment (Python 3.13, macOS 14.4.1).

```bash
pip list | grep transcrypt
Using Python 3.13.1 environment at: venv313
transcrypt              3.9.3
```

Here's the Python code

```python
from sweetbean.stimulus import Text
from sweetbean import Block, Experiment

introduction_stimulus = Text(text="Welcome to our awesome experiment.")
introduction_list = [introduction_stimulus]
introduction_block = Block(introduction_list)
block_list = [introduction_block]
experiment = Experiment(block_list)
experiment.to_html("index.html")
```

Here's the generated index.html. Does this index.html run work locally for you?

```html
<!DOCTYPE html>
<head>
<title>My awesome experiment</title><script src="https://unpkg.com/jspsych@7.3.1"></script>
<script src="https://unpkg.com/@jspsych/plugin-html-keyboard-response@1.1.2"></script>
<script src="https://unpkg.com/@jspsych/plugin-survey-text@1.1.2"></script>
<script src="https://unpkg.com/@jspsych/plugin-survey-multi-choice@1.1.2"></script>
<script src="https://unpkg.com/@jspsych/plugin-survey-likert@1.1.2"></script>
<script src="https://unpkg.com/@jspsych/plugin-survey-likert@1.1.2"></script>
<link href="https://unpkg.com/jspsych@7.3.1/css/jspsych.css" rel="stylesheet" type="text/css"/>
<script src="https://unpkg.com/@jspsych/plugin-image-keyboard-response@1.1.2"></script>
<script src="https://unpkg.com/@jspsych/plugin-video-keyboard-response@1.1.2"></script>
<script src="https://unpkg.com/@jspsych-contrib/plugin-rok@1.1.1"></script>
<script src="https://unpkg.com/@jspsych-contrib/plugin-html-choice@1.0.0"></script>
<script src="https://cdn.jsdelivr.net/npm/sweetbean@0.0.6/dist/runtime-script.js"></script>
<link href="https://cdn.jsdelivr.net/npm/sweetbean@0.0.6/dist/style/main.css" rel="stylesheet" type="text/css"/>
<link href="https://cdn.jsdelivr.net/npm/sweetbean@0.0.6/dist/style/bandit.css" rel="stylesheet" type="text/css"/>
</head>
<body></body>
<script>
jsPsych = initJsPsych();
trials = [
{timeline: [{type: jsPsychHtmlKeyboardResponse,trial_duration:()=>{let trial_duration=null;return trial_duration},stimulus:()=>{let stimulus=(None)('white', 'Welcome to our awesome experiment.');return stimulus},choices:()=>{let choices=[];return choices},correct_key:()=>{let correct_key='';return correct_key},on_finish:(data)=>{let trial_duration=null;data["bean_trial_duration"]=trial_duration;let duration=null;data["bean_duration"]=duration;let stimulus=(None)('white', 'Welcome to our awesome experiment.');data["bean_stimulus"]=stimulus;let choices=[];data["bean_choices"]=choices;let correct_key='';data["bean_correct_key"]=correct_key;let type='jsPsychHtmlKeyboardResponse';data["bean_type"]=type;let text='Welcome to our awesome experiment.';data["bean_text"]=text;let color='white';data["bean_color"]=color;data["bean_correct"]=data["bean_correct_key"]===data["response"];}}], timeline_variables: []}]
;jsPsych.run(trials)</script>
</html>
```
