### 2025-03-13 1a24357
add error checking
