+++ .github/workflows/test-pytest-1.yml
@@ -1,29 +1,0 @@
-# This workflow will install Python dependencies, run tests and lint with a variety of Python versions
-# For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
-
-name: Test with py.test
-
-on:
-  pull_request:
-  merge_group:
-  workflow_call:
-  push:
-    branches:
-      - main
-
-jobs:
-  test:
-    strategy:
-      fail-fast: true
-      matrix:
-        python-version: ["3.9", "3.10", "3.11", "3.12"]
-        os: [ubuntu-latest, macos-14, windows-latest]
-    runs-on: ${{ matrix.os }}
-    steps:
-    - uses: actions/checkout@v3
-    - uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python-version }}
-        cache: "pip"
-    - run: pip install ".[test]"
-    - run: pytest tests/
