### 2025-03-13 AutoResearch/sweetbean#127 @younesStrittmatter
Hi @hauselin,

thanks a lot for your patience!

I am sorry this is quite complicated to debug since on my machine I can't replicate the error (I am using the exact same python and sweetbean version).

Anyways, I have a suspicion where the behavior stems from and changed that to be hopefully more robust. Could you try the newest version of sweetbean that is published as pre-release? (`v1.1.3a`)

Sorry again for this. I hope the change fix the problem or at least throw an error when there is a problem.

(the index.hml file created by your code doesn't work for me either)
