### 2025-03-14 AutoResearch/sweetbean#127 @hauselin
After installing the latest version 1.1.3a0, I'm getting a new error, which I'm not sure was intended or not. I could download java and try fix it, but not sure if that's what you were expecting?

```python

from sweetbean.stimulus import Text
from sweetbean import Block, Experiment

introduction_stimulus = Text(text="Welcome to our awesome experiment.")
introduction_list = [introduction_stimulus]
introduction_block = Block(introduction_list)
block_list = [introduction_block]
experiment = Experiment(block_list)
experiment.to_html("index.html")

The operation couldn’t be completed. Unable to locate a Java Runtime.
Please visit http://www.java.com/ for information on installing Java.

The operation couldn’t be completed. Unable to locate a Java Runtime.
Please visit http://www.java.com/ for information on installing Java.

The operation couldn’t be completed. Unable to locate a Java Runtime.
Please visit http://www.java.com/ for information on installing Java.

The operation couldn’t be completed. Unable to locate a Java Runtime.
Please visit http://www.java.com/ for information on installing Java.

The operation couldn’t be completed. Unable to locate a Java Runtime.
Please visit http://www.java.com/ for information on installing Java.

The operation couldn’t be completed. Unable to locate a Java Runtime.
Please visit http://www.java.com/ for information on installing Java.
```
