### 2025-03-15 AutoResearch/sweetbean#127 @younesStrittmatter
This is super helpful. Thank you so much!
I think you are right and java is indeed a dependency. Can you try installing it and see if it solves the problem?

If so, I will add the dependency to the documentation and check for it on import of the library to point the user to the missing dependency.
