### 2025-03-15 6fc65f7
add java dependency
