### 2025-03-17 AutoResearch/sweetbean#127 @hauselin
Yes, it works after installing java. Please highlight/update this dependency in the docs. Thanks!
