### 2025-03-17 ba377aa
add java test to init
