### 2025-03-17 34529bd
relax java dependency to 17
