+++ src/sweetbean/__init__.py
@@ -7,14 +7,14 @@ from sweetbean.experiment import Experiment
 
 
 def check_java():
-    """Ensure that Java 19+ is installed."""
+    """Ensure that Java 17+ is installed."""
     java_path = shutil.which("java")
     if java_path is None:
         sys.stderr.write(
             "\n❌ Error: Java is required but not found.\n"
-            "   Please install Java 19+ using one of the following:\n"
-            "   - macOS: brew install openjdk@19\n"
-            "   - Ubuntu/Debian: sudo apt install openjdk-19-jdk\n"
+            "   Please install Java 17+ using one of the following:\n"
+            "   - macOS: brew install openjdk@17\n"
+            "   - Ubuntu/Debian: sudo apt install openjdk-17-jdk\n"
             "   - Windows: Download Java from https://jdk.java.net/archive/\n"
         )
         sys.exit(1)
@@ -32,9 +32,9 @@ def check_java():
         version = version_line.split('"')[1]
         major_version = int(version.split(".")[0])
 
-        if major_version < 19:
+        if major_version < 17:
             sys.stderr.write(
-                f"\n❌ Error: Java 19+ is required. Found Java {major_version}.\n"
+                f"\n❌ Error: Java 17+ is required. Found Java {major_version}.\n"
             )
             sys.exit(1)
 
