### 2025-03-17 71eb353
relax java dependency to java being installed
