### 2025-03-17 AutoResearch/sweetbean#127 @younesStrittmatter
Added it to the docs and there is a warning now when importing the package without java being installed. Sorry for the inconvenience and thanks again a lot!
