### 2025-03-19 #7703 @sappelhoff
@younesStrittmatter  please let me know once you finished your items from the list above: https://github.com/openjournals/joss-reviews/issues/7703#issuecomment-2735688836

I will then work on my final items, and then it's (hopefully) time to ship it!
