+++ paper/paper.bib
@@ -1,11 +1,17 @@
+
+
 @article{de2023jspsych,
-  title={jsPsych: Enabling an open-source collaborative ecosystem of behavioral experiments},
-  author={de Leeuw, Joshua R and Gilbert, Rebecca A and Luchterhandt, Bj{\"o}rn},
-  journal={Journal of Open Source Software},
-  volume={8},
-  number={85},
-  pages={5351},
-  year={2023}
+title = {jsPsych: Enabling an Open-Source Collaborative Ecosystem of Behavioral Experiments},
+doi = {10.21105/joss.05351},
+url = {https://doi.org/10.21105/joss.05351},
+year = {2023},
+publisher = {The Open Journal},
+volume = {8},
+number = {85},
+pages = {5351},
+author = {Joshua R. de Leeuw and Rebecca A. Gilbert and Björn Luchterhandt},
+
+journal = {Journal of Open Source Software}
 }
 
 @article{autora2024,
