### 2025-03-20 #7703 @younesStrittmatter
Hi, I've finished the items on the list and also added the missing DOI.

Here is the zenodo DOI:
10.5281/zenodo.15054706
