### 2025-03-20 #7703 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
